chore: create Container Connectio action component
