+++ packages/renderer/src/lib/preferences/PreferencesResourcesRendering.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-import { faArrowUpRightFromSquare, faPlay, faRotateRight, faStop, faTrash } from '@fortawesome/free-solid-svg-icons';
+import { faArrowUpRightFromSquare } from '@fortawesome/free-solid-svg-icons';
 import Fa from 'svelte-fa/src/fa.svelte';
 import { providerInfos } from '../../stores/providers';
 import type { ProviderContainerConnectionInfo, ProviderInfo } from '../../../../main/src/plugin/api/provider-info';
@@ -13,20 +13,15 @@ import { filesize } from 'filesize';
 import { router } from 'tinro';
 import SettingsPage from './SettingsPage.svelte';
 import ConnectionStatus from '../ui/ConnectionStatus.svelte';
-import { ConnectionCallback, eventCollect, startTask } from './preferences-connection-rendering-task';
-import type { IContainerStatus } from './Util';
-import LoadingIconButton from '../ui/LoadingIconButton.svelte';
+import { eventCollect } from './preferences-connection-rendering-task';
+import { getContainerConnectionName, IContainerRestart, IContainerStatus } from './Util';
+import { Buffer } from 'buffer';
+import PreferencesContainerConnectionActions from './PreferencesContainerConnectionActions.svelte';
 
 interface IProviderContainerConfigurationPropertyRecorded extends IConfigurationPropertyRecordedSchema {
   value?: any;
   container: string;
   providerId: string;
-}
-
-interface IContainerRestart {
-  provider: string;
-  container: string;
-  loggerHandlerKey: symbol;
 }
 
 export let properties: IConfigurationPropertyRecordedSchema[] = [];
@@ -36,7 +31,7 @@ $: containerConnectionStatus = new Map<string, IContainerStatus>();
 let isStatusUpdated = false;
 
 let configurationKeys: IConfigurationPropertyRecordedSchema[];
-let containersRestarting: IContainerRestart[] = [];
+let restartingQueue: IContainerRestart[] = [];
 
 let providersUnsubscribe: Unsubscriber;
 let configurationPropertiesUnsubscribe: Unsubscriber;
@@ -82,9 +77,9 @@ onMount(() => {
 });
 
 function getContainerRestarting(provider: string, container: string): IContainerRestart {
-  const containerToRestart = containersRestarting.filter(c => c.provider === provider && c.container === container)[0];
+  const containerToRestart = restartingQueue.filter(c => c.provider === provider && c.container === container)[0];
   if (containerToRestart) {
-    containersRestarting = containersRestarting.filter(c => c.provider !== provider && c.container !== container);
+    restartingQueue = restartingQueue.filter(c => c.provider !== provider && c.container !== container);
   }
   return containerToRestart;
 }
@@ -136,153 +131,45 @@ $: providerContainerConfiguration = tmpProviderContainerConfiguration
     return map;
   }, new Map<string, IProviderContainerConfigurationPropertyRecorded[]>());
 
-function getLoggerHandler(
+function updateContainerStatus(
   provider: ProviderInfo,
   containerConnectionInfo: ProviderContainerConnectionInfo,
-): ConnectionCallback {
-  return {
-    log: () => {},
-    warn: () => {},
-    error: args => {
-      setContainerStatusUpdateFailed(provider, containerConnectionInfo, args);
-    },
-    onEnd: () => {},
-  };
+  action?: string,
+  error?: string,
+): void {
+  const containerConnectionName = getContainerConnectionName(provider, containerConnectionInfo);
+  if (error) {
+    const currentStatus = containerConnectionStatus.get(containerConnectionName);
+    containerConnectionStatus.set(containerConnectionName, {
+      ...currentStatus,
+      inProgress: false,
+      error,
+    });
+  } else if (action) {
+    containerConnectionStatus.set(containerConnectionName, {
+      inProgress: true,
+      action: action,
+      status: containerConnectionInfo.status,
+    });
+  }
+  containerConnectionStatus = containerConnectionStatus;
+}
+
+function addContainerToRestartingQueue(container: IContainerRestart) {
+  restartingQueue.push(container);
 }
 
 async function startContainerProvider(
   provider: ProviderInfo,
   containerConnectionInfo: ProviderContainerConnectionInfo,
-  loggerHandlerKey?: symbol,
-): Promise<void> {
-  try {
-    if (containerConnectionInfo.status === 'stopped') {
-      if (!loggerHandlerKey) {
-        setContainerStatusIsChanging(provider, 'start', containerConnectionInfo);
-        loggerHandlerKey = startTask(
-          `Start ${provider.name} ${containerConnectionInfo.name}`,
-          `/preferences/resources`,
-          getLoggerHandler(provider, containerConnectionInfo),
-        );
-      }
-      await window.startProviderConnectionLifecycle(
-        provider.internalId,
-        containerConnectionInfo,
-        loggerHandlerKey,
-        eventCollect,
-      );
-    }
-  } catch (e) {
-    console.error(e);
-  }
-}
-
-async function stopContainerProvider(
-  provider: ProviderInfo,
-  containerConnectionInfo: ProviderContainerConnectionInfo,
-): Promise<void> {
-  try {
-    if (containerConnectionInfo.status === 'started') {
-      setContainerStatusIsChanging(provider, 'stop', containerConnectionInfo);
-      const loggerHandlerKey = startTask(
-        `Stop ${provider.name} ${containerConnectionInfo.name}`,
-        `/preferences/resources`,
-        getLoggerHandler(provider, containerConnectionInfo),
-      );
-      await window.stopProviderConnectionLifecycle(
-        provider.internalId,
-        containerConnectionInfo,
-        loggerHandlerKey,
-        eventCollect,
-      );
-    }
-  } catch (e) {
-    console.error(e);
-  }
-}
-
-async function restartContainerProvider(
-  provider: ProviderInfo,
-  containerConnectionInfo: ProviderContainerConnectionInfo,
-): Promise<void> {
-  if (containerConnectionInfo.status === 'started') {
-    setContainerStatusIsChanging(provider, 'restart', containerConnectionInfo);
-    const loggerHandlerKey = startTask(
-      `Restart ${provider.name} ${containerConnectionInfo.name}`,
-      `/preferences/resources`,
-      getLoggerHandler(provider, containerConnectionInfo),
-    );
-    containersRestarting.push({
-      container: containerConnectionInfo.name,
-      provider: provider.internalId,
-      loggerHandlerKey,
-    });
-    await window.stopProviderConnectionLifecycle(
-      provider.internalId,
-      containerConnectionInfo,
-      loggerHandlerKey,
-      eventCollect,
-    );
-  }
-}
-
-async function deleteContainerProvider(
-  provider: ProviderInfo,
-  containerConnectionInfo: ProviderContainerConnectionInfo,
-): Promise<void> {
-  try {
-    if (containerConnectionInfo.status === 'stopped' || containerConnectionInfo.status === 'unknown') {
-      setContainerStatusIsChanging(provider, 'delete', containerConnectionInfo);
-      const loggerHandlerKey = startTask(
-        `Delete ${provider.name} ${containerConnectionInfo.name}`,
-        `/preferences/resources`,
-        getLoggerHandler(provider, containerConnectionInfo),
-      );
-      await window.deleteProviderConnectionLifecycle(
-        provider.internalId,
-        containerConnectionInfo,
-        loggerHandlerKey,
-        eventCollect,
-      );
-    }
-  } catch (e) {
-    console.error(e);
-  }
-}
-
-function setContainerStatusIsChanging(
-  provider: ProviderInfo,
-  action: string,
-  containerConnectionInfo: ProviderContainerConnectionInfo,
-): void {
-  containerConnectionStatus.set(getContainerConnectionName(provider, containerConnectionInfo), {
-    inProgress: true,
-    action: action,
-    status: containerConnectionInfo.status,
-  });
-  containerConnectionStatus = containerConnectionStatus;
-}
-
-function setContainerStatusUpdateFailed(
-  provider: ProviderInfo,
-  containerConnectionInfo: ProviderContainerConnectionInfo,
-  error: string,
-): void {
-  const containerConnectionName = getContainerConnectionName(provider, containerConnectionInfo);
-  const currentStatus = containerConnectionStatus.get(containerConnectionName);
-  containerConnectionStatus.set(containerConnectionName, {
-    ...currentStatus,
-    inProgress: false,
-    error,
-  });
-  containerConnectionStatus = containerConnectionStatus;
-}
-
-function getContainerConnectionName(
-  provider: ProviderInfo,
-  containerConnectionInfo: ProviderContainerConnectionInfo,
-): string {
-  return `${provider.name}-${containerConnectionInfo.name}`;
+  loggerHandlerKey: symbol,
+) {
+  await window.startProviderConnectionLifecycle(
+    provider.internalId,
+    containerConnectionInfo,
+    loggerHandlerKey,
+    eventCollect,
+  );
 }
 </script>
 
@@ -330,7 +217,15 @@ function getContainerConnectionName(
           {#each provider.containerConnections as container}
             <div class="px-5 py-2 w-[240px]">
               <div class="float-right text-gray-700 cursor-not-allowed">
-                <Fa icon="{faArrowUpRightFromSquare}" />
+                <button
+                  on:click="{() =>
+                    router.goto(
+                      `/preferences/container-connection/${provider.internalId}/${Buffer.from(
+                        container.endpoint.socketPath,
+                      ).toString('base64')}`,
+                    )}">
+                  <Fa icon="{faArrowUpRightFromSquare}" />
+                </button>
               </div>
               <div class="{container.status !== 'started' ? 'text-gray-500' : ''} text-sm">
                 {container.name}
@@ -368,52 +263,12 @@ function getContainerConnectionName(
                   {/each}
                 </div>
               {/if}
-              {#if containerConnectionStatus.has(getContainerConnectionName(provider, container))}
-                {@const state = containerConnectionStatus.get(getContainerConnectionName(provider, container))}
-                {#if container.lifecycleMethods && container.lifecycleMethods.length > 0}
-                  <div class="mt-2 relative">
-                    <!-- TODO: see action available like machine infos -->
-                    <div class="flex bg-zinc-900 w-fit rounded-lg m-auto">
-                      {#if container.lifecycleMethods.includes('start')}
-                        <div class="ml-2">
-                          <LoadingIconButton
-                            clickAction="{() => startContainerProvider(provider, container)}"
-                            action="start"
-                            icon="{faPlay}"
-                            state="{state}"
-                            leftPosition="left-[0.15rem]" />
-                        </div>
-                      {/if}
-                      {#if container.lifecycleMethods.includes('start') && container.lifecycleMethods.includes('stop')}
-                        <LoadingIconButton
-                          clickAction="{() => restartContainerProvider(provider, container)}"
-                          action="restart"
-                          icon="{faRotateRight}"
-                          state="{state}"
-                          leftPosition="left-1.5" />
-                      {/if}
-                      {#if container.lifecycleMethods.includes('stop')}
-                        <LoadingIconButton
-                          clickAction="{() => stopContainerProvider(provider, container)}"
-                          action="stop"
-                          icon="{faStop}"
-                          state="{state}"
-                          leftPosition="left-[0.22rem]" />
-                      {/if}
-                      {#if container.lifecycleMethods.includes('delete')}
-                        <div class="mr-2 text-sm">
-                          <LoadingIconButton
-                            clickAction="{() => deleteContainerProvider(provider, container)}"
-                            action="delete"
-                            icon="{faTrash}"
-                            state="{state}"
-                            leftPosition="left-1" />
-                        </div>
-                      {/if}
-                    </div>
-                  </div>
-                {/if}
-              {/if}
+              <PreferencesContainerConnectionActions
+                provider="{provider}"
+                container="{container}"
+                containerConnectionStatuses="{containerConnectionStatus}"
+                updateContainerStatus="{updateContainerStatus}"
+                addContainerToRestartingQueue="{addContainerToRestartingQueue}" />
               <div class="mt-1.5 text-gray-500 text-[9px]">
                 <div>{provider.name} {provider.version ? `v${provider.version}` : ''}</div>
               </div>
@@ -421,6 +276,17 @@ function getContainerConnectionName(
           {/each}
           {#each provider.kubernetesConnections as kubeConnection}
             <div class="px-5 py-2 w-[240px]">
+              <div class="float-right text-gray-700 cursor-not-allowed">
+                <button
+                  on:click="{() =>
+                    router.goto(
+                      `/preferences/kubernetes-connection/${provider.internalId}/${Buffer.from(
+                        kubeConnection.endpoint.apiURL,
+                      ).toString('base64')}`,
+                    )}">
+                  <Fa icon="{faArrowUpRightFromSquare}" />
+                </button>
+              </div>
               <div class="text-sm">
                 {kubeConnection.name}
               </div>
